+++ src/Controller/WardrobeItemController.php
@@ -11,6 +11,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpClient\HttpClient;
 
 #[Route('/wardrobe/item')]
 final class WardrobeItemController extends AbstractController
@@ -178,5 +180,81 @@ final class WardrobeItemController extends AbstractController
             'wardrobe_item' => $wardrobeItem,
             'recommended_items' => $recommendations,
         ]);
+    }
+
+    #[Route('/ai/analyze', name: 'app_ai_analyze', methods: ['POST'])]
+    public function analyze(Request $request): JsonResponse
+    {
+        // Récupérer le fichier image depuis le formulaire
+        $uploadedFile = $request->files->get('image');
+        if (!$uploadedFile) {
+            return new JsonResponse(['error' => 'Aucun fichier image fourni.'], 400);
+        }
+
+        // Lire le contenu du fichier
+        $imageContent = file_get_contents($uploadedFile->getPathname());
+        $imageData = base64_encode($imageContent);
+        $sampleImageData = substr($imageData, 0, 100) . '...';
+
+        // Construire le prompt pour l'API OpenAI
+        $prompt = "Here is a base64 snippet of an image of a clothing item: $sampleImageData\n\n" .
+          "Please analyze this image and extract the following details:\n" .
+          "1. name: The name or style of the garment.\n" .
+          "2. brand: The brand or designer of the garment.\n" .
+          "3. color: The main color of the garment.\n" .
+          "4. description: A brief description of the garment’s style and features.\n\n" .
+          "Return your answer as a valid JSON object with the keys \"name\", \"brand\", \"color\", and \"description\". " .
+          "If any detail cannot be determined, use an empty string for that key.";
+
+        // Créer un client HTTP
+        $client = HttpClient::create();
+
+        // Récupérer la clé API depuis l'environnement
+        $apiKey = trim($_ENV['OPENAI_API_KEY_WARDROBE']);
+
+        if (!$apiKey) {
+            return new JsonResponse(['error' => 'Clé API non configurée.'], 500);
+        }
+
+        try {
+            // Appeler l'API OpenAI
+            $response = $client->request('POST', 'https://api.openai.com/v1/chat/completions', [
+                'headers' => [
+                    'Authorization' => 'Bearer ' . $apiKey,
+                    'Content-Type'  => 'application/json',
+                ],
+                'json' => [
+                    'model' => 'gpt-3.5-turbo-0125', // ou "gpt-4" si vous y avez accès
+                    'messages' => [
+                        [
+                            'role' => 'system',
+                            'content' => 'You are an expert in clothing analysis. Extract garment details from image data.'
+                        ],
+                        [
+                            'role' => 'user',
+                            'content' => $prompt,
+                        ]
+                    ],
+                    'temperature' => 0.7,
+                ]
+            ]);
+
+            $result = $response->toArray();
+            $contentString = $result['choices'][0]['message']['content'] ?? '';
+            // Nettoyer le contenu si nécessaire (par exemple, retirer les délimiteurs Markdown)
+            $cleanContent = preg_replace('/```(json)?/', '', $contentString);
+            $cleanContent = trim($cleanContent);
+
+            $parsedOutput = json_decode($cleanContent, true);
+            if (json_last_error() !== JSON_ERROR_NONE) {
+                return new JsonResponse([
+                    'error' => 'Erreur lors du parsing JSON: ' . json_last_error_msg(),
+                    'raw_output' => $cleanContent
+                ], 500);
+            }
+            return new JsonResponse($parsedOutput);
+        } catch (\Exception $e) {
+            return new JsonResponse(['error' => $e->getMessage()], 500);
+        }
     }
 }
